+++ axual-webclient/applications.go
@@ -52,11 +52,13 @@ func (c *Client) CreateApplication(data ApplicationRequest) (*ApplicationRespons
 	return &o, nil
 }
 
-
-func (c *Client) GetApplicationsByAttributes(attributes url.Values) (*ApplicationsByAttributesResponse, error) {
-    o := ApplicationsByAttributesResponse{}
-
-    url := fmt.Sprintf("%s/applications/search/findByAttributes?%s", c.ApiURL, attributes.Encode())
+func (c *Client) GetApplicationByNameOrShortName(params url.Values) (*ApplicationResponse, error) {
+    o := ApplicationResponse{}
+	endpoint := fmt.Sprintf("findByName?name=%s", params.Get("name")) 
+	if params.Get("shortName") != "" {
+    	endpoint = fmt.Sprintf("findByShortName?shortName=%s", params.Get("shortName"))
+	}
+    url := fmt.Sprintf("%s/applications/search/%s", c.ApiURL, endpoint)
 	fmt.Println("URL", url)
     err := c.RequestAndMap("GET", url, nil, nil, &o)
     if err != nil {