+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 * Added support for searching by shortName in Instance, Environment, Application Data Sources.
 * Removed trial guide
 * Added Support for importing Application Deployment
+* Retrieve an application using `findByName` or `findByShortName` endpoints instead of `findByAttributes`
 
 ## [2.5.5](https://github.com/Axual/terraform-provider-axual/releases/tag/v2.5.5) - 2025-03-07
 * Removed unused method for `/groups/{uid}/members/{uid}`